ShowsPopularAdapter: extract SearchResultDiffCallback.
