+++ app/src/main/java/com/battlelancer/seriesguide/ui/search/SearchResultDiffCallback.kt
@@ -1,0 +1,18 @@
+package com.battlelancer.seriesguide.ui.search
+
+import androidx.recyclerview.widget.DiffUtil
+
+/** [DiffUtil.ItemCallback] for [SearchResult] items for use with a RecyclerView adapter. */
+class SearchResultDiffCallback : DiffUtil.ItemCallback<SearchResult>() {
+
+    override fun areItemsTheSame(oldItem: SearchResult, newItem: SearchResult): Boolean =
+        oldItem.tvdbid == newItem.tvdbid && oldItem.title == newItem.title
+
+    override fun areContentsTheSame(oldItem: SearchResult, newItem: SearchResult): Boolean {
+        return (oldItem.state == newItem.state
+                && oldItem.language == newItem.language
+                && oldItem.posterPath == newItem.posterPath
+                && oldItem.overview == newItem.overview)
+    }
+
+}
