+++ app/src/main/java/com/battlelancer/seriesguide/ui/search/ShowsPopularAdapter.kt
@@ -2,11 +2,13 @@ package com.battlelancer.seriesguide.ui.search
 
 import android.view.ViewGroup
 import androidx.paging.PagedListAdapter
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 
-class ShowsPopularAdapter(val onItemClickListener: AddFragment.AddAdapter.OnItemClickListener) :
-        PagedListAdapter<SearchResult, RecyclerView.ViewHolder>(DIFF_CALLBACK) {
+class ShowsPopularAdapter(
+    val onItemClickListener: AddFragment.AddAdapter.OnItemClickListener
+) : PagedListAdapter<SearchResult, RecyclerView.ViewHolder>(
+    SearchResultDiffCallback()
+) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
         return ShowsPopularViewHolder.create(parent, onItemClickListener)
@@ -43,20 +45,6 @@ class ShowsPopularAdapter(val onItemClickListener: AddFragment.AddAdapter.OnItem
             }
         }
         notifyDataSetChanged()
-    }
-
-    companion object {
-        private val DIFF_CALLBACK = object : DiffUtil.ItemCallback<SearchResult>() {
-            override fun areItemsTheSame(oldItem: SearchResult, newItem: SearchResult): Boolean =
-                    oldItem.tvdbid == newItem.tvdbid && oldItem.title == newItem.title
-
-            override fun areContentsTheSame(oldItem: SearchResult, newItem: SearchResult): Boolean {
-                return (oldItem.state == newItem.state
-                        && oldItem.language == newItem.language
-                        && oldItem.posterPath == newItem.posterPath
-                        && oldItem.overview == newItem.overview)
-            }
-        }
     }
 
 }
